+++ tests/strategies/test_encoding.py
@@ -23,6 +23,7 @@ from tilings.strategies import (
     EmptyCellInferralFactory,
     FactorFactory,
     FactorInsertionFactory,
+    FusableRowAndColumnPlacementFactory,
     FusionFactory,
     InsertionEncodingVerificationStrategy,
     LocallyFactorableVerificationStrategy,
@@ -475,6 +476,7 @@ strategy_objects = (
     + [
         AssumptionPointingFactory(),
         DummyStrategy(),
+        FusableRowAndColumnPlacementFactory(),
         PointingStrategy(),
         UnfusionRowStrategy(),
         UnfusionColumnStrategy(),